Scrapers/LinkedInScraper/requirements.txt
selenium>=4.15.0
webdriver-manager>=4.0.0
selenium-stealth>=1.0.6
beautifulsoup4>=4.12.0
pandas>=2.1.0
openpyxl>=3.1.0
python-dotenv>=1.0.0
